refactor(ui): renamed TUI to ui
TUI: Text-based user interface or Touristik Union International
UI: User interface
