+++ src/include/ui/helper_functions.h
@@ -1,0 +1,5 @@
+#ifndef UI_HELPER_FUNCTIONS_H
+#define UI_HELPER_FUNCTIONS_H
+void m_is_prime(void);
+void m_get_prime(void);
+#endif // UI_HELPER_FUNCTIONS_H
