+++ src/include/ui/rsa.h
@@ -1,0 +1,7 @@
+#ifndef UI_RSA_H
+#define UI_RSA_H
+
+void m_rsa(void);
+void m_rsa_encrypt(void);
+
+#endif // UI_RSA_H
